fix(mobile): resolve touch issues and unify button styling in plugins settings
